+++ packages/lesswrong/server.ts
@@ -68,6 +68,7 @@ import './server/updateScores';
 import './server/siteAdminMetadata';
 import './server/callbacks';
 import './server/notificationCallbacks';
+import './server/notificationCallbacksHelpers';
 import './server/voteServer';
 import './server/recommendations';
 import './server/emails/emailTokens';
