+++ packages/lesswrong/server/notificationCallbacks.tests.ts
@@ -1,41 +1,112 @@
 import { testStartup } from "../testing/testMain";
-// import { createDummyPost } from "../testing/utils";
-// import { getSubscribedUsers, postsNewNotifications } from "./notificationCallbacks";
+import { createDummyPost, createDummyLocalgroup } from "../testing/utils";
+import { postsNewNotifications } from "./notificationCallbacks";
+import { createNotifications } from "./notificationCallbacksHelpers";
 
 testStartup()
 
-jest.mock('./notificationCallbacks', () => {
-  const originalModule = jest.requireActual('./notificationCallbacks');
+jest.mock('./notificationCallbacksHelpers', () => {
+  const originalModule = jest.requireActual('./notificationCallbacksHelpers');
 
-  // we just want to mock getSubscribedUsers()
+  // control which users are associated with which notifications (for testing postsNewNotifications)
+  const mockGetSubscribedUsers = jest.fn().mockImplementation((document) => {
+    // return users subscribed to the local group
+    if (document.collectionName === 'Localgroups') return [{_id: '111'}]
+    // return users subscribed to the author
+    if (document.collectionName === 'Users') return [{_id: '222'}, {_id: '333'}]
+  })
+  
   return {
     __esModule: true,
     ...originalModule,
-    getSubscribedUsers: jest.fn(),
+    getSubscribedUsers: mockGetSubscribedUsers,
+    getUsersWhereLocationIsInNotificationRadius: jest.fn().mockResolvedValue([{_id: '222'}]),
+    createNotifications: jest.fn()
   };
 })
- 
 
-describe("postsNewNotifications", () => {
-  it("creates the correct type of notification", () => {
-    // There is a bug with this logic, where postsNewNotifications() can notify users for the wrong reason,
-    // which means it may notify them in a way they did not want.
-    //
-    // For example, take a user who wants on-site notifications for a specific author's posts and emails for
-    // posts in their local group. If that author creates a new post in a group *that is not that local group*,
-    // we will send the user an email about it, rather than an on-site notification. However, the user would expect
-    // to get an on-site notification and not an email.
-    //
-    // I feel like the fix may require significant refactoring of postsNewNotifications(), since a post may
-    // qualify for multiple notification types
-    // (ex. if the author above posted in the user's local group, we may want to send both notification types)
-    // and we also need to be sure not to send multiple of any notification.
-    //
-    // JP and I (Sarah) intended to write tests around this issue, but since I don't know how impactful this is
-    // and I don't think I have time to properly fix this now, I will leave this for the next person.
-    
+beforeEach(() => {
+  jest.clearAllMocks();
+});
 
-    // const testPost = createDummyPost()
-    // await postsNewNotifications(testPost)
+describe("test postsNewNotifications", () => {
+  it("only sends the newPost notifications when the new post is not in a group and not an event", async () => {
+    const testPost = await createDummyPost()
+    await postsNewNotifications(testPost)
+    // notify both users subscribed to the author
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['222', '333'],
+      notificationType: 'newPost',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+  })
+  
+  it("sends the newEventInRadius and newPost notifications when the new post is an event, not in a group", async () => {
+    const testPost = await createDummyPost()
+    testPost.isEvent = true
+    testPost.mongoLocation = true
+    await postsNewNotifications(testPost)
+    // only send one notification per user
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['222'],
+      notificationType: 'newEventInRadius',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['333'],
+      notificationType: 'newPost',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+  })
+  
+  it("sends the newGroupPost and newPost notifications when the new post is in a group, not an event", async () => {
+    const testPost = await createDummyPost()
+    const testGroup = await createDummyLocalgroup()
+    testPost.groupId = testGroup._id
+    await postsNewNotifications(testPost)
+
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['111'],
+      notificationType: 'newGroupPost',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['222', '333'],
+      notificationType: 'newPost',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+  })
+  
+  it("sends the newEvent, newEventInRadius, and newPost notifications when the new post is an event in a group", async () => {
+    const testPost = await createDummyPost()
+    const testGroup = await createDummyLocalgroup()
+    testPost.groupId = testGroup._id
+    testPost.isEvent = true
+    testPost.mongoLocation = true
+    await postsNewNotifications(testPost)
+    // only send one notification per user
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['111'],
+      notificationType: 'newEvent',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['222'],
+      notificationType: 'newEventInRadius',
+      documentType: 'post',
+      documentId: testPost._id
+    })
+    expect(createNotifications).toHaveBeenCalledWith({
+      userIds: ['333'],
+      notificationType: 'newPost',
+      documentType: 'post',
+      documentId: testPost._id
+    })
   })
 })
